- replace generator with newer implementation
